docs: add template draft and remove tooltip from select template
